Finished up adding the rest of the fields to the new recipe controller and page and the submit and cancel buttons.
Added the appropriate unit tests for the AddOrUpdateRecipeController. Basically tests for the cookbook and tests for the promisse returned by the resource.$save call.

Deleted the GivenAUserResource test file, as it was just going to test the resource. I still wanted to test how the resource is defined, but for now, couldn't figure out how to inject the $resource into the factory for the rbUser resource.